Split community queries
