+++ client/coral-admin/src/routes/Community/containers/CommunityMenu.js
@@ -1,30 +1,4 @@
-import { compose, gql } from 'react-apollo';
-import { mapProps } from 'recompose';
-import withQuery from 'coral-framework/hocs/withQuery';
+import withCachedFlaggedUsernamesCount from '../hocs/withCachedFlaggedUsernamesCount';
 import CommunityMenu from '../components/CommunityMenu';
-import get from 'lodash/get';
 
-const withData = withQuery(
-  gql`
-    query TalkAdmin_CommunityMenu {
-      flaggedUsernamesCount: userCount(
-        query: {
-          action_type: FLAG
-          state: { status: { username: [SET, CHANGED] } }
-        }
-      )
-    }
-  `,
-  {
-    options: {
-      fetchPolicy: 'cache-only',
-    },
-  }
-);
-
-export default compose(
-  withData,
-  mapProps(ownProps => ({
-    flaggedUsernamesCount: get(ownProps, 'root.flaggedUsernamesCount'),
-  }))
-)(CommunityMenu);
+export default withCachedFlaggedUsernamesCount(CommunityMenu);
