+++ client/coral-admin/src/routes/Community/containers/FlaggedUser.js
@@ -4,17 +4,17 @@ import { withFragments } from 'plugin-api/beta/client/hocs';
 
 export default withFragments({
   root: gql`
-    fragment TalkAdminCommunity_FlaggedUser_root on RootQuery {
+    fragment TalkAdmin_Community_FlaggedUser_root on RootQuery {
       __typename
     }
   `,
   me: gql`
-    fragment TalkAdminCommunity_FlaggedUser_me on User {
+    fragment TalkAdmin_Community_FlaggedUser_me on User {
       id
     }
   `,
   user: gql`
-    fragment TalkAdminCommunity_FlaggedUser_user on User {
+    fragment TalkAdmin_Community_FlaggedUser_user on User {
       id
       username
       created_at
